+++ src/models/mod.rs
@@ -1,6 +1,7 @@
 pub mod bookmark;
 pub mod folder;
 pub mod qbt;
+pub mod recovery_key;
 pub mod settings;
 pub mod user;
 pub mod vault;
@@ -8,6 +9,7 @@ pub mod vault;
 pub use bookmark::Bookmark;
 pub use folder::IndexedFolder;
 pub use qbt::QbTorrent;
+pub use recovery_key::RecoveryKey;
 pub use settings::Settings;
 pub use user::{NewUser, User};
 pub use vault::{EncryptedFile, VaultConfig};